Fix violation of SA1402 by ColorPaneConverters.cs
