+++ PowerPointLabs/PowerPointLabs/Converters/ColorPane/HSLColorToRGBColor.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Drawing;
+using System.Windows.Data;
+
+namespace PowerPointLabs.Converters.ColorPane
+{
+    class HSLColorToRGBColor : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            Color selectedColor = (HSLColor)value;
+            return Color.FromArgb(255,
+                    selectedColor.R,
+                    selectedColor.G,
+                    selectedColor.B);
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
