+++ PowerPointLabs/PowerPointLabs/Converters/ColorPane/IsActiveBoolToButtonBackColorConverter.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Drawing;
+using System.Windows.Data;
+
+namespace PowerPointLabs.Converters.ColorPane
+{
+    class IsActiveBoolToButtonBackColorConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            bool isActive = (bool)value;
+            if (isActive)
+            {
+                return SystemColors.ActiveCaption;
+            }
+            return SystemColors.Control;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
