+++ PowerPointLabs/PowerPointLabs/Converters/ColorPane/SelectedColorToAnalogousHigher.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Windows.Data;
+using PowerPointLabs.ColorPicker;
+
+namespace PowerPointLabs.Converters.ColorPane
+{
+    class SelectedColorToAnalogousHigher : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            var selectedColor = (HSLColor)value;
+            return ColorHelper.GetColorShiftedByAngle(selectedColor, 30.0f);
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
