--- PowerPointLabs/PowerPointLabs/Converters/ColorPaneConverters.cs
@@ -1,334 +1,0 @@
-﻿using System;
-using System.Drawing;
-using System.Windows.Data;
-using PowerPointLabs.ColorPicker;
-
-namespace PowerPointLabs.Converters
-{
-    class HSLColorToRGBColor : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            Color selectedColor = (HSLColor) value;
-            return Color.FromArgb(255,
-                    selectedColor.R,
-                    selectedColor.G,
-                    selectedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToAnalogousLower : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, -30.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToAnalogousHigher : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 30.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToComplementaryColor : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 180.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToSplitComplementaryLower : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 150.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToSplitComplementaryHigher : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 210.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToTriadicLower : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, -120.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToTriadicHigher : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 120.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToTetradicOne : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 90.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToTetradicTwo : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 180.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    class SelectedColorToTetradicThree : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return ColorHelper.GetColorShiftedByAngle(selectedColor, 270.0f);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticOne : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor) value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.80f*240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticTwo : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor) value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.70f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticThree : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.60f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticFour : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.50f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticFive : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.40f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticSix : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.30f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToMonochromaticSeven : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, 0.20f * 240);
-            return Color.FromArgb(255,
-                convertedColor.R,
-                convertedColor.G,
-                convertedColor.B);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class SelectedColorToBrightnessValue : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return (int)(selectedColor.Luminosity);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-    class SelectedColorToSaturationValue : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            var selectedColor = (HSLColor)value;
-            return (int)(selectedColor.Saturation);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    class IsActiveBoolToButtonBackColorConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            bool isActive = (bool)value;
-            if (isActive)
-            {
-                return SystemColors.ActiveCaption;
-            }
-            return SystemColors.Control;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
